FTR tsfy combobox service (#35493)
* migrate combobox to typescript
